requirements.txt
streamlit>=1.36
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
faiss-cpu>=1.8
pypdf>=4.2
tiktoken
python-dotenv
python-docx
docx2txt